It's fixed but messy
Both players go until the whole board is filled. The fix is messy becuase it's it's just if statements repeated for both players. After the whole board is filled, the Tic Tac Toe board is just printed over and over.

Do to: add a way for the game to know when it's done and who won.